+++ core/src/main/java/com/graphhopper/reader/osm/WaySegmentParser.java
@@ -62,7 +62,6 @@ public class WaySegmentParser {
     private static final Logger LOGGER = LoggerFactory.getLogger(WaySegmentParser.class);
     private static final Set<String> INCLUDE_IF_NODE_TAGS = new HashSet<>(Arrays.asList("barrier", "highway", "railway", "crossing", "ford"));
 
-    private ToDoubleFunction<ReaderNode> elevationProvider = node -> 0d;
     private Predicate<ReaderWay> wayFilter = way -> true;
     private Predicate<ReaderNode> splitNodeFilter = node -> false;
     private WayPreprocessor wayPreprocessor = (way, coordinateSupplier, nodeTagSupplier) -> {
@@ -109,7 +108,8 @@ public class WaySegmentParser {
         LOGGER.info("Finished reading OSM file." +
                 " pass1: " + (int) sw1.getSeconds() + "s, " +
                 " pass2: " + (int) sw2.getSeconds() + "s, " +
-                " total: " + (int) (sw1.getSeconds() + sw2.getSeconds()) + "s");
+                " total: " + (int) (sw1.getSeconds() + sw2.getSeconds()) + "s" +
+                " memory: " + Helper.getMemInfo());
     }
 
     /**
@@ -203,7 +203,7 @@ public class WaySegmentParser {
                 LOGGER.info("pass2 - processed nodes: " + nf(nodeCounter) + ", accepted nodes: " + nf(acceptedNodes) +
                         ", " + Helper.getMemInfo());
 
-            long nodeType = nodeData.addCoordinatesIfMapped(node.getId(), node.getLat(), node.getLon(), () -> elevationProvider.applyAsDouble(node));
+            long nodeType = nodeData.addCoordinatesIfMapped(node.getId(), node.getLat(), node.getLon());
             if (nodeType == EMPTY_NODE)
                 return;
 
@@ -415,14 +415,6 @@ public class WaySegmentParser {
          */
         public Builder(PointAccess pointAccess, Directory directory) {
             waySegmentParser = new WaySegmentParser(new OSMNodeData(pointAccess, directory));
-        }
-
-        /**
-         * @param elevationProvider used to determine the elevation of an OSM node
-         */
-        public Builder setElevationProvider(ToDoubleFunction<ReaderNode> elevationProvider) {
-            waySegmentParser.elevationProvider = elevationProvider;
-            return this;
         }
 
         /**