+++ core/src/test/java/com/graphhopper/reader/dem/CGIARProviderTest.java
@@ -44,6 +44,7 @@ public class CGIARProviderTest {
     @BeforeEach
     public void setUp() {
         instance = new CGIARProvider(tempDir.toString());
+        instance.init();
     }
 
     @AfterEach