+++ core/src/test/java/com/graphhopper/routing/RoutingAlgorithmWithOSMTest.java
@@ -111,7 +111,7 @@ public class RoutingAlgorithmWithOSMTest {
     public void testMonacoMotorcycleCurvature() {
         List<Query> queries = new ArrayList<>();
         queries.add(new Query(43.730729, 7.42135, 43.727697, 7.419199, 2675, 117));
-        queries.add(new Query(43.727687, 7.418737, 43.74958, 7.436566, 3727, 170));
+        queries.add(new Query(43.727687, 7.418737, 43.74958, 7.436566, 3730, 170));
         queries.add(new Query(43.728677, 7.41016, 43.739213, 7.4277, 2769, 167));
         queries.add(new Query(43.733802, 7.413433, 43.739662, 7.424355, 2373, 137));
         queries.add(new Query(43.730949, 7.412338, 43.739643, 7.424542, 2203, 116));
@@ -265,12 +265,12 @@ public class RoutingAlgorithmWithOSMTest {
     public void testMonacoFoot3D() {
         // most routes have same number of points as testMonaceFoot results but longer distance due to elevation difference
         List<Query> queries = createMonacoFoot();
-        queries.get(0).getPoints().get(1).expectedDistance = 1627;
+        queries.get(0).getPoints().get(1).expectedDistance = 1624;
         queries.get(2).getPoints().get(1).expectedDistance = 2250;
         queries.get(3).getPoints().get(1).expectedDistance = 1482;
 
         // or slightly longer tour with less nodes: list.get(1).setDistance(1, 3610);
-        queries.get(1).getPoints().get(1).expectedDistance = 3573;
+        queries.get(1).getPoints().get(1).expectedDistance = 3576;
         queries.get(1).getPoints().get(1).expectedPoints = 149;
 
         Profile profile = TestProfiles.accessSpeedAndPriority("foot");
@@ -318,9 +318,9 @@ public class RoutingAlgorithmWithOSMTest {
     public void testMonacoBike3D() {
         List<Query> queries = new ArrayList<>();
         // 1. alternative: go over steps 'Rampe Major' => 1.7km vs. around 2.7km
-        queries.add(new Query(43.730864, 7.420771, 43.727687, 7.418737, 2702, 111));
+        queries.add(new Query(43.730864, 7.420771, 43.727687, 7.418737, 2670, 118));
         // 2.
-        queries.add(new Query(43.728499, 7.417907, 43.74958, 7.436566, 4220, 233));
+        queries.add(new Query(43.728499, 7.417907, 43.74958, 7.436566, 4223, 233));
         // 3.
         queries.add(new Query(43.728677, 7.41016, 43.739213, 7.427806, 2776, 167));
         // 4.
@@ -329,7 +329,7 @@ public class RoutingAlgorithmWithOSMTest {
         // try reverse direction
         // 1.
         queries.add(new Query(43.727687, 7.418737, 43.730864, 7.420771, 2598, 115));
-        queries.add(new Query(43.74958, 7.436566, 43.728499, 7.417907, 4250, 165));
+        queries.add(new Query(43.74958, 7.436566, 43.728499, 7.417907, 3982, 181));
         queries.add(new Query(43.739213, 7.427806, 43.728677, 7.41016, 2806, 145));
         // 4. avoid tunnel(s)!
         queries.add(new Query(43.739662, 7.424355, 43.733802, 7.413433, 1901, 116));
@@ -337,6 +337,7 @@ public class RoutingAlgorithmWithOSMTest {
         GraphHopper hopper = createHopper(MONACO,
                 new Profile("bike").setCustomModel(CustomModel.merge(getCustomModel("bike.json"), getCustomModel("bike_elevation.json")).
                         addToPriority(If("!bike_access", MULTIPLY, "0"))));
+        hopper.setEncodedValuesString("average_slope, max_slope, " + hopper.getEncodedValuesString());
         hopper.setElevationProvider(new SRTMProvider(DIR));
         hopper.importOrLoad();
         checkQueries(hopper, queries);
@@ -590,6 +591,7 @@ public class RoutingAlgorithmWithOSMTest {
 
         GraphHopper hopper = createHopper(BAYREUTH, new Profile("bike").setCustomModel(
                 CustomModel.merge(getCustomModel("bike.json"), getCustomModel("bike_elevation.json"))));
+        hopper.setEncodedValuesString("average_slope, max_slope, " + hopper.getEncodedValuesString());
         hopper.setElevationProvider(new SRTMProvider(DIR));
         hopper.importOrLoad();
         checkQueries(hopper, list);
@@ -705,7 +707,7 @@ public class RoutingAlgorithmWithOSMTest {
                 setStoreOnFlush(false).
                 setOSMFile(osmFile).
                 setProfiles(profiles).
-                setEncodedValuesString("average_slope, max_slope, hike_rating, car_access, car_average_speed, " +
+                setEncodedValuesString("hike_rating, car_access, car_average_speed, " +
                         "foot_access, foot_priority, foot_average_speed, foot_network, " +
                         "bike_access, bike_priority, bike_average_speed, bike_network, roundabout, " +
                         "mtb_access, mtb_priority, mtb_average_speed, mtb_rating, " +